Minor changes (rename)
