+++ app/depends/__init__.py
@@ -2,6 +2,6 @@ from typing import Annotated
 
 from fastapi import Depends
 
-from app.manager import PipelineSession, get_pipeline_session
+from app.session import PipelineSession, get_pipeline_session
 
 DependPipelineSession = Annotated[PipelineSession, Depends(get_pipeline_session)]
